fixed warnings

+++ source/gui/blob.h
@@ -3,9 +3,9 @@
 
 #include "quickjs.h"
 
-size_t JS_GetBlobSize(JSContext *ctx, JSValueConst blob);
+int64_t JS_GetBlobSize(JSContext *ctx, JSValueConst blob);
 JSValue JS_GetBlobParts(JSContext *ctx, JSValueConst blob_obj);
 
-size_t JS_GetBlobUint8Array(JSContext *ctx, JSValueConst blob, uint8_t *out_buf);
+int64_t JS_GetBlobUint8Array(JSContext *ctx, JSValueConst blob, uint8_t *out_buf);
 
 #endif
